// src/app/interfaces/medium/tellink/alarm-profile.ts
export interface AlarmProfile
{
    id: number;
    name: string;
    contract_id: number;
    contract_name: string;
    date: string;
    status_flag: number;
    on_rele1: number;
    off_rele1: number;
    on_rele2: number;
    off_rele2: number;
    switch_on: number;
    switch_off: number;
    low_GSM_signal: number;
    no_GPRS_data: number;
    meter_error: number;
    input1: number;
    input2: number;
    input3: number;
    input4: number;
    max_power_rele_off: number;
    max_power_rele_on: number;
    min_power_factor: number;
    min_current_rele_on_phase1: number;
    min_current_rele_on_phase2: number;
    min_current_rele_on_phase3: number;
    max_current_rele_on_phase1: number;
    max_current_rele_on_phase2: number;
    max_current_rele_on_phase3: number;
    max_current_rele_off_phase1: number;
    max_current_rele_off_phase2: number;
    max_current_rele_off_phase3: number;
    min_voltage_phase1: number;
    min_voltage_phase2: number;
    min_voltage_phase3: number;
    max_voltage_phase1: number;
    max_voltage_phase2: number;
    max_voltage_phase3: number;
    max_power_value_rele_off: string;
    max_power_value_rele_on: string;
    power_factor: string;
    max_current_value_rele_on_phase1: string;
    max_current_value_rele_on_phase2: string;
    max_current_value_rele_on_phase3: string;
    min_current_value_rele_on_phase1: string;
    min_current_value_rele_on_phase2: string;
    min_current_value_rele_on_phase3: string;
    max_voltage_value: string;
    min_voltage_value: string;
    max_current_value_rele_off_phase1: string;
    max_current_value_rele_off_phase2: string;
    max_current_value_rele_off_phase3: string;
    warning_mail_1: string;
    warning_mail_2: string;
    warning_mail_3: string;
    input5: number;
    input6: number;
    input7: number;
    input8: number;
    on_rele3: number;
    off_rele3: number;
    gps_error: number;
}